Splitting out some utility functions
