+++ src/XML/src/FMI3/fmi3_xml_parser.c
@@ -203,32 +203,6 @@ void fmi3_xml_parse_free_context(fmi3_xml_parser_context_t* context) {
     context->callbacks->free(context);
 }
 
-void fmi3_xml_parse_fatal(fmi3_xml_parser_context_t* context, const char* fmt, ...) {
-    va_list args;
-    va_start (args, fmt);
-    jm_log_fatal_v(context->callbacks, module, fmt, args);
-    va_end (args);
-    XML_StopParser(context->parser,0);
-}
-
-void fmi3_xml_parse_error(fmi3_xml_parser_context_t* context, const char* fmt, ...) {
-    va_list args;
-    va_start (args, fmt);
-    if (context->parser)
-        jm_log_info(context->callbacks, module, "[Line:%u] Detected during parsing:", XML_GetCurrentLineNumber(context->parser));
-    jm_log_error_v(context->callbacks, module, fmt, args);
-    va_end (args);
-}
-
-void fmi3_xml_parse_warning(fmi3_xml_parser_context_t* context, const char* fmt, ...) {
-    va_list args;
-    va_start (args, fmt);
-    if (context->parser)
-        jm_log_info(context->callbacks, module, "[Line:%u] Detected during parsing:", XML_GetCurrentLineNumber(context->parser));
-    jm_log_warning_v(context->callbacks, module, fmt, args);
-    va_end (args);
-}
-
 /**
  * Raises a generic parse error for the given attribute.
  */
