+++ src/XML/src/FMI3/fmi3_xml_parser.h
@@ -30,6 +30,7 @@
 #include "fmi3_xml_model_description_parser.h"
 #include "fmi3_xml_terminals_and_icons_parser.h"
 #include "fmi3_xml_parser_context_impl.h"
+#include "fmi3_xml_parser_util.h"
 
 #ifdef __cplusplus
 extern "C" {
@@ -85,10 +86,6 @@ jm_vector(char) * fmi3_xml_reserve_parse_buffer(fmi3_xml_parser_context_t* conte
 int fmi3_xml_alloc_parse_buffer(fmi3_xml_parser_context_t* context, size_t items);
 void fmi3_xml_free_variable_start_values(fmi3_xml_parser_context_t* context);
 void fmi3_xml_free_parse_buffer(fmi3_xml_parser_context_t* context);
-
-void fmi3_xml_parse_fatal  (fmi3_xml_parser_context_t* context, const char* fmt, ...);
-void fmi3_xml_parse_error  (fmi3_xml_parser_context_t* context, const char* fmt, ...);
-void fmi3_xml_parse_warning(fmi3_xml_parser_context_t* context, const char* fmt, ...);
 
 /**
  * Raises generic parse error for given attribute.
